Fix off by one error in MatrixPlot
The previous calculations for x- and ylimits / 'pixel' positions were
incorrect. Considering `MatrixPlot(rand(2, 2), (1, 2), (1, 2))` the
previous code takes the given x- and yrange as limits, which means one
obtains a plot showing the x- and y-intervals (1, 2) having a length of
1. However, the step size is 1 and thus the 'pixels' have a 'width' and
'height' of 1. But the plot shows only the area for (1, 2)x(1, 2) which
fits one but not all four 'pixels'. In the general case this leads to
the first column and last row of a given Matrix not to be shown.

Now every entry of the Matrix is shown and the ranges specified by
x/yrange are no longer aligned with the center of the 'pixels' at the
border, instead e.g. xrange[1] is now aligned with the left border of
the 'pixels' produced by the first column of the given matrix.